+++ tests/test_edge_cases.py
@@ -659,6 +659,49 @@ from tests.test_main import DATA_DIR
                 'please combine it with the docstring of the class'
             ],
         ),
+        (
+            '35_doc404_return_and_yield_double_unwrap/google.py',
+            {
+                'style': 'google',
+                'argTypeHintsInDocstring': False,
+                'checkYieldTypes': True,
+                'checkReturnTypes': True,
+            },
+            [
+                'DOC404: Function `iteratorDictMismatch` yield type(s) in docstring '
+                'not consistent with the return annotation. The yield type (the 0th '
+                'arg in Generator[...]/Iterator[...]): Dict[str, Any]; docstring '
+                '"yields" section types: str'
+            ],
+        ),
+        (
+            '35_doc404_return_and_yield_double_unwrap/numpy.py',
+            {
+                'style': 'numpy',
+                'checkYieldTypes': True,
+                'checkReturnTypes': True,
+            },
+            [
+                'DOC404: Function `iteratorDictMismatch` yield type(s) in docstring '
+                'not consistent with the return annotation. The yield type (the 0th '
+                'arg in Generator[...]/Iterator[...]): Dict[str, Any]; docstring '
+                '"yields" section types: str'
+            ],
+        ),
+        (
+            '35_doc404_return_and_yield_double_unwrap/sphinx.py',
+            {
+                'style': 'sphinx',
+                'checkYieldTypes': True,
+                'checkReturnTypes': True,
+            },
+            [
+                'DOC404: Function `iteratorDictMismatch` yield type(s) in docstring '
+                'not consistent with the return annotation. The yield type (the 0th '
+                'arg in Generator[...]/Iterator[...]): Dict[str, Any]; docstring '
+                '"yields" section types: str'
+            ],
+        ),
     ],
 )
 def testEdgeCases(